Added user authorization for edit/delete methods

+++ backend/routes/posts.js
@@ -38,7 +38,8 @@ multer({storage: storage}).single("image"),
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then((createdPost) => {
     res.status(201).json({
@@ -57,7 +58,6 @@ router.put(
   (req, res, next) => {
     let imagePath = req.body.imagePath;
     if(req.file) {
-      console.log(req.file);
       const url = req.protocol + '://' + req.get('host');
       imagePath = url + '/images/' + req.file.filename;
     }
@@ -65,10 +65,15 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
-    Post.updateOne({_id: req.params.id}, post).then(result => {
-      console.log(result);
-      res.status(200).json({message: 'updated'});
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+      if(result.nModified > 0){
+        res.status(200).json({message: 'updated'});
+      } else {
+        res.status(401).json({message: 'Not Authorized'});
+      }
     });
   }
 );
@@ -112,9 +117,12 @@ router.delete(
   '/:id',
   checkAuth,
   (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    console.log(result);
-    res.status(200).json({message: "post deleted"});
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+    if(result.n > 0){
+      res.status(200).json({message: 'Deletion Successful'});
+    } else {
+      res.status(401).json({message: 'Not Authorized'});
+    }
   });
 });
